ai-executive-studio/services/orchestrator/pyproject.toml
[tool.poetry]
name = "ai-executive-orchestrator"
version = "0.1.0"
description = "Orchestrator service for AI Executive Studio"
authors = ["AI Executive Studio <dev@ai-exec.local>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.110.0"
uvicorn = { extras = ["standard"], version = "0.27.1" }
structlog = "23.2.0"
python-dotenv = "1.0.1"
pyjwt = "2.8.0"
httpx = "0.27.0"
langchain = "0.1.13"
llama-index = "0.10.18"
qdrant-client = "1.8.2"
aiokafka = "0.10.0"
pyyaml = "6.0.1"
langfuse = "2.0.9"
opentelemetry-sdk = "1.24.0"
opentelemetry-api = "1.24.0"
opentelemetry-exporter-otlp = "1.24.0"
opentelemetry-instrumentation-fastapi = "0.45b0"
prometheus-fastapi-instrumentator = "6.1.0"
python-multipart = "0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "8.1.1"
pytest-asyncio = "0.23.5"
httpx = "0.27.0"
ruff = "0.3.3"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
